pricing: revamp the /subscription page
